feat: Rename executable
Why:
Makes it a little more generic, less specific to SVD, and keeps an
option open for other config formats

feat: Name enums structs and unions

Why:
In some IDEs, the help text would only show that a register in the
peripheral was an ananymous struct or similar.  Naming the structs, etc.
allowed that IDE to display the doc comments.